Refactoring : split main file in 4 smaller files.
